chore: rename no_rules.rs to no_category.rs
Self-explanatory naming. 'no_rules' is cryptic; 'no_category' matches the VITEST-NO-* rule category pattern.
