+++ src/rules/no_category.rs
